:sparkles: 로그인 및 로그아웃 기능 연결
